Implement management of "signal vector"
"Signal vector" is an array of integers having values 0 or 1,
which represent presense or absence of an OS signal
with the particular number on the system the paclage was
compiled for.
The signal with number 1 (SIGHUP) is at index 0;
the signal with the maximum number is at the last index,
which is unknown until the run time.