Simplify management of commands bound to events
Greatly simplified management of commands bound to events
and provided (prospective) robustness for cross-thread accesses
as the command objects are now duplicated on both reading
and writing if the object is about to cross the thread boundary.

NOTE that appending to the existing thread using the "+"
prefix does not work anymore (to be reimplemented).