Block all signals when executing our handler
It turns out, that if we do not specify blocking of
all signals when installing our signal handler,
a deadlock can occur: if calls to two or more signal
handlers in the same thread overlap, they all will
end up blocked on the same mutex protecting the access
to the table of syncpoints.
More correct solution would possibly be implementing
a per-syncpoint mutex, but currently we just make
each thread serve at most one signal at a time.