Support offsets for [info sigrtmin/sigrtmax]
"sigrtmin" and "sigrtmax" [info] topics now support
one optional integer argument which, if present
is interpreted as an offset to add to the signum
to which SIGRTMIN or SIGRTMAX pthrads macros
resolve at the time of a call.
This can be used to conveniently refer to the real-time
signals as they cannot be referred to by names,
only by base signum + offset.