Modify syncpoints to use signal map
The syncpoints management code now uses
the new generic signal map container.

NOTE that [posix::send] currently crashes
in the syncpoints table traversal code as the
syncpoints are not properly initialized.
The solution possibly requires implementing
something like Tcl hash "entries" so that
the CreateSigMapEntry() returns the "entry"
which can then be used to get and/or set the
actual value managed by that entry.