Add support for "client data" in syncpoints
This is a preliminary attempt to attach saved
sigaction info to syncpoints.
The problem is that currently a syncpoint is "acquired"
before the signal handler is installed (and the old
sigaction info for that signal becomes available),
and the process of syncpoint acquisition is a one-stop
operation which combines possible creation of the syncpoint
with the initialization of its values.
It seems we need to expose more syncpoint manipulation API
so that the [trap] code could have "getter" and "setter"
procedures to manipulate the syncpoints.